Moved functions from strain_tools to plotting_tools.
